fix error tester

+++ srcs/manage_list/get_lower_idx.c
@@ -6,7 +6,7 @@
 /*   By: mdouglas <mdouglas@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/12/22 14:29:21 by mdouglas          #+#    #+#             */
-/*   Updated: 2022/12/22 17:32:12 by mdouglas         ###   ########.fr       */
+/*   Updated: 2022/12/25 14:40:59 by mdouglas         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,10 +21,10 @@
  */
 size_t  get_lower_index(t_chain *stack)
 {
-    int     first_node;
-    int     index;
-    int     count;
-    t_node  *helper;
+    int         first_node;
+    size_t      index;
+    size_t      count;
+    t_node      *helper;
 
     helper = stack->head;
     index = 0;
